toolkit/components/extensions/test/mochitest/test_extension_contentscript.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test for content script</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SpawnTask.js"></script>
  <script type="text/javascript;version=1.8" src="head.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>

<script type="application/javascript;version=1.8">
"use strict";

const BASE = "http://mochi.test:8888/tests/toolkit/components/extensions/test/mochitest";

add_task(function* test_contentscript()
{
  let extension = loadExtension("content_script");
  yield extension.startup();
  info("extension loaded");

  let loadingCount = 0;
  let interactiveCount = 0;
  let completeCount = 0;
  extension.onMessage("script-run-loading", () => { loadingCount++; });
  extension.onMessage("script-run-interactive", () => { interactiveCount++; });

  let completePromise = new Promise(resolve => {
    extension.onMessage("script-run-complete", () => { completeCount++; resolve(); });
  });

  yield new Promise(resolve => { setTimeout(resolve, 0); });

  let win = window.open();

  win.location = "file_contentscript_page1.html";

  ok(true, "page loaded");

  yield Promise.all([waitForLoad(win), completePromise]);
  info("test page loaded");

  win.close();

  is(loadingCount, 1, "document_start script ran exactly once");
  is(interactiveCount, 1, "document_end script ran exactly once");
  is(completeCount, 1, "document_idle script ran exactly once");

  yield extension.unload();
  info("extension unloaded");
});

function waitForLoad(win) {
  return new Promise(resolve => {
    win.addEventListener("load", function listener() {
      win.removeEventListener("load", listener, true);
      resolve();
    }, true);
  });
}
</script>

</body>
</html>
